Add assignment generation functionality in backend and frontend
The new update introduces a feature to generate bulk assignments in the application. Changes are made in both back-end and front-end files. In the back-end, a function is implemented in the assignment controller for generating assignments. For the front-end, an assignments service function for calling this API and a new page component for invoking this service have been introduced along with some route changes.