Implement dark/light theme switch and migrate CSS to SCSS
The primary modification in this commit is the implementation of a dark/light theme switch functionality, enabled by the creation of a new theme service. This led to the migration of all CSS styles to SCSS to benefit from improved organisation and modularity. The style URLs in all components have been updated, and FontAwesome has been added to project dependencies to provide the relevant icons for the switch button.